add tester for host level tests
This makes the transition from using Mininet containers to using
the entire machine, thus eliminating one layer of virtualization.

Traffic shaping is performed using tc commands. The manner in which
Mininet was using tc was suboptimal; if a latency was specified, it
would use the TBF qdisc, despite the request for HTB. Moreover, we
can now test with the HFSC qdisc as well thanks to the new '--scheduler'
parameter.

Given the extensive rewriting and the completely different context
in which this version operates, the version has been bumped to version
4, codename Christina.